refactor(backend): バックエンドコードのリファクタリング
- main.pyから設定、モデル、サービス、ユーティリティを分離
- config.py: 環境変数と定数の管理を分離
- models.py: Pydanticモデルを独立したモジュールに移動
- services/: Google Mapsサービスとルートサービスを分離
- utils/: ジオメトリ関連のユーティリティ関数を分離
- main.pyを簡潔化し、エンドポイント定義に集中
